feat: refactored code created new file for classifier definition
